more code smells

+++ cpp/coefficients.cpp
@@ -42,7 +42,7 @@ dolfinx_contact::pack_coefficient_quadrature(
     throw std::invalid_argument("Unsupported integral type.");
   }
   // Create quadrature rule
-  QuadratureRule q_rule(cell_type, q_degree, entity_dim);
+  QuadratureRule q_rule(cell_type, q_degree, (int)entity_dim);
 
   // Get element information
   const dolfinx::fem::FiniteElement* element
